+++ packages/core/src/debugUtilities/debugMode.ts
@@ -1,10 +1,14 @@
+import logger from '../logger';
+
 let debugMode = false;
 
 export const enableDebugMode = () => {
+    logger.setLevel('debug');
     debugMode = true;
 };
 
 export const disableDebugMode = () => {
+    logger.resetLevel();
     debugMode = false;
 };